Move SessionKeys from Invitation
