+++ src/Simplex/RemoteControl/Invitation.hs
@@ -8,30 +8,26 @@
 
 module Simplex.RemoteControl.Invitation where
 
-import Control.Concurrent.STM (TVar)
 import Control.Monad (unless)
-import Crypto.Random (ChaChaDRG)
 import Data.Aeson.TH (deriveJSON)
 import qualified Data.Attoparsec.ByteString.Char8 as A
 import Data.ByteString (ByteString)
 import qualified Data.ByteString as B
-import Data.List.NonEmpty (NonEmpty (..))
 import Data.Maybe (isJust)
 import Data.Text (Text)
 import Data.Text.Encoding (decodeUtf8Lenient, encodeUtf8)
-import Data.Time.Clock.System (SystemTime, getSystemTime)
+import Data.Time.Clock.System (SystemTime)
 import Data.Word (Word16, Word32)
 import Network.HTTP.Types (parseSimpleQuery)
 import Network.HTTP.Types.URI (SimpleQuery, renderQuery, urlDecode)
-import qualified Network.TLS as TLS
 import qualified Simplex.Messaging.Crypto as C
-import Simplex.Messaging.Crypto.SNTRUP761.Bindings (KEMPublicKey, KEMSecretKey, sntrup761Keypair)
+import Simplex.Messaging.Crypto.SNTRUP761.Bindings (KEMPublicKey)
 import Simplex.Messaging.Encoding (Encoding (..))
 import Simplex.Messaging.Encoding.String
 import Simplex.Messaging.Parsers (defaultJSON)
 import Simplex.Messaging.Transport.Client (TransportHost)
-import Simplex.Messaging.Transport.Credentials (genCredentials, tlsCredentials)
 import Simplex.Messaging.Version (VersionRange, mkVersionRange)
+import Simplex.RemoteControl.Types
 
 data XRCPInvitation = XRCPInvitation
   { -- | CA TLS certificate fingerprint of the controller.
@@ -185,29 +181,6 @@ instance Encoding XRCPEncryptedInvitation where
     pure XRCPEncryptedInvitation {dhPubKey, encryptedInvitation}
 
 -- * Utils
-
--- | A bunch of keys that should be generated by a controller to start a new remote session and produce invites
-data CtrlSessionKeys = CtrlSessionKeys
-  { ts :: SystemTime,
-    ca :: C.KeyHash,
-    credentials :: TLS.Credentials,
-    sSigKey :: C.PrivateKeyEd25519,
-    dhKey :: C.PrivateKeyX25519,
-    kem :: (KEMPublicKey, KEMSecretKey)
-  }
-
-newCtrlSessionKeys :: TVar ChaChaDRG -> (C.APrivateSignKey, C.SignedCertificate) -> IO CtrlSessionKeys
-newCtrlSessionKeys rng (caKey, caCert) = do
-  ts <- getSystemTime
-  (_, C.APrivateDhKey C.SX25519 dhKey) <- C.generateDhKeyPair C.SX25519
-  (_, C.APrivateSignKey C.SEd25519 sSigKey) <- C.generateSignatureKeyPair C.SEd25519
-
-  let parent = (C.signatureKeyPair caKey, caCert)
-  sessionCreds <- genCredentials (Just parent) (0, 24) "Session"
-  let (ca, credentials) = tlsCredentials $ sessionCreds :| [parent]
-  kem <- sntrup761Keypair rng
-
-  pure CtrlSessionKeys {ts, ca, credentials, sSigKey, dhKey, kem}
 
 sessionXRCPInvitation ::
   -- | App information
